refactor(cloister): extract boot reconciliation apply
