+++ scripts/file-size-baseline.txt
@@ -16,7 +16,6 @@
 1013 src/dashboard/server/routes/workspaces/workspace-data.ts
 1520 src/dashboard/server/services/issue-data-service.ts
 1306 src/dashboard/server/ws-rpc.ts
-1052 src/lib/cloister/deacon-auto-resume.ts
 1148 src/lib/cloister/deacon-swarm.ts
 3631 src/lib/cloister/deacon.ts
 1407 src/lib/cloister/merge-agent.ts
